Moved common stuff to RemoteStorage
